// apps/frontend/src/api/model/assignmentsResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * My Catatan Kuliah
 * My Catatan Kuliah API
 * OpenAPI spec version: 1.0.0
 */
import type { AssignmentsAssignmentStatus } from './assignmentsAssignmentStatus';

export interface AssignmentsResponse {
  readonly id: string;
  course_id: string;
  title: string;
  status: AssignmentsAssignmentStatus;
  position: number;
  due_date: string;
  color?: string;
  readonly created_by: string;
  readonly created_at: string;
  readonly updated_at: string;
}